jobready/requirements.txt
django==5.1.2
django-rest-framework==0.1.0
axios==0.4.0
django-cors-headers==4.5.0
gunicorn==23.0.0
pillow==11.0.0
environs==11.0.0
whitenoise==6.7.0
uvicorn==0.32.0
typing_extensions==4.12.2
joblib==1.4.2
dj-database-url==2.2.0